Cells,Crypto,Network,Tests: add support for NTor handshake
CreateFast can only be used for first hop of a circuit and only
If we don't have fingerprint and onion key for any router
(e.g. during bootstraping and before we have any consensus data)
so we need to add support for nTor or legacy TAP (Useful for backward
compatiblity with older routers) or both to be able to create
multi-hop circuits.