FSharpLint: enable PascalCase config for public elements
Even though F# docs recommend officially camelCase [1], we feel this
rule goes against the historical recommendations that Microsoft always
had for public APIs in .NET (which are still respected in C#).

An F# purist would say that both requirements can also be respected
via using an attribute on the camelCased F# function that exposes it
with a PascalCase API: the CompiledName attribute. But it's easy to
forget to use this attribute everywhere (and there's no FSharpLint rule
for it yet) so we will stick with PascalCase as our CONTRIBUTING guide
already recommends.

[1] https://stackoverflow.com/a/57525412/544947