Network: set time limit for circuit operations like Create and Extend
In some cases, create or extend operations might not result
in any error (maybe an inaccessable node?), in these instances
we need to be able to throw if the operation takes too long.

This commit also removes test-level timeout for multihop circuit
because create and extend have timeout themselves now.