Client,Network,Services: add TorClient
This commit aims to introduce a simple API for end users.

Working with network especially Tor routers is flaky and
delegating retry logic implementation to users causes NOnion to
be unreliable in CI and in normal use, this commit introduces
retry logic in some places where it's needed the most. This
should hopefully make NOnion more reliable.