Tests: upgrade to .NET 6
This commit upgrades test project and our CI to use .NET 6.
There's no excuse to use .netcoreapp 3.1 anymore.

It's noteworthy to say that we will keep NOnion target platform
on netstandard2.0 for the time being.

P.S:
My reason for this upgrade was mostly lack of native apple
silicon support in .NET Core 3.1.